Factor out GridFilterModal and change EntityFieldFilterModal to extend it
