+++ src/org/labkey/test/components/ui/search/EntityFieldFilterModal.java
@@ -4,7 +4,7 @@ import org.labkey.test.BootstrapLocators;
 import org.labkey.test.Locator;
 import org.labkey.test.components.UpdatingComponent;
 import org.labkey.test.components.bootstrap.ModalDialog;
-import org.labkey.test.components.react.Tabs;
+import org.labkey.test.components.ui.grids.GridFilterModal;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
@@ -12,28 +12,17 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 /**
  * Wraps 'labkey-ui-component' defined in <code>internal/components/search/EntityFieldFilterModal.tsx</code>
  */
-public class EntityFieldFilterModal extends ModalDialog
+public class EntityFieldFilterModal extends GridFilterModal
 {
-    private static final Locator listItem = Locator.byClass("list-group-item");
-
-    private final UpdatingComponent _linkedComponent;
-
     public EntityFieldFilterModal(WebDriver driver, UpdatingComponent linkedComponent)
     {
-        super(new ModalDialog.ModalDialogFinder(driver));
-        _linkedComponent = linkedComponent;
+        super(driver, linkedComponent);
     }
 
     @Override
     protected void waitForReady(ModalDialog.ElementCache ec)
     {
-        super.waitForReady(ec);
-
-        // TODO need to factor out the QueryFilterPanel part of this modal
-        getWrapper().shortWait().until(wb ->
-            listItem.findElementOrNull(elementCache().querySelectionPanel) != null ||
-            listItem.findElementOrNull(elementCache().fieldsSelectionPanel) != null
-        );
+        getWrapper().shortWait().until(ExpectedConditions.visibilityOf(listItem.findElementOrNull(elementCache().querySelectionPanel)));
     }
 
     /**
@@ -53,21 +42,6 @@ public class EntityFieldFilterModal extends ModalDialog
     }
 
     /**
-     * Select field to configure filters for
-     * @param fieldLabel Field's label
-     * @return this component
-     */
-    public EntityFieldFilterModal selectField(String fieldLabel)
-    {
-        WebElement fieldItem = listItem.withText(fieldLabel).findElement(elementCache().fieldsSelectionPanel);
-        fieldItem.click();
-        Locator.byClass("filter-modal__col-sub-title").withText("Find values for " + fieldLabel)
-                .waitForElement(elementCache().filterPanel, 10_000);
-
-        return this;
-    }
-
-    /**
      * Select parent and field to configure filters for
      * @param parentName name of parent type
      * @param fieldLabel Field's label
@@ -79,48 +53,6 @@ public class EntityFieldFilterModal extends ModalDialog
         selectField(fieldLabel);
 
         return this;
-    }
-
-    /**
-     * Select the filter expression tab for the current field.
-     * @return panel for configuring the filter expression
-     */
-    public FilterExpressionPanel selectExpressionTab()
-    {
-        elementCache().filterTabs.selectTab("Filter");
-        return elementCache().filterExpressionPanel;
-    }
-
-    /**
-     * Select the facet tab for the current field. Will throw <code>NoSuchElementException</code> if tab isn't present.
-     * @return panel for configuring faceted filter
-     */
-    public FilterFacetedPanel selectFacetTab()
-    {
-        elementCache().filterTabs.selectTab("Choose values");
-        return elementCache().filterFacetedPanel;
-    }
-
-    /**
-     * Save current changes to the search criteria.
-     * Throw <code>IllegalStateException</code> if the save button is disabled because no changes have been made.
-     */
-    public void confirm()
-    {
-        if (!elementCache().submitButton.isEnabled())
-        {
-            throw new IllegalStateException("Confirmation button is not enabled.");
-        }
-
-        _linkedComponent.doAndWaitForUpdate(() -> {
-            elementCache().submitButton.click();
-            waitForClose();
-        });
-    }
-
-    public void cancel()
-    {
-        dismiss("Cancel");
     }
 
     @Override
@@ -135,26 +67,11 @@ public class EntityFieldFilterModal extends ModalDialog
         return new ElementCache();
     }
 
-    protected class ElementCache extends ModalDialog.ElementCache
+    protected class ElementCache extends GridFilterModal.ElementCache
     {
         // Entities column
         protected final WebElement querySelectionPanel = Locator.byClass("filter-modal__col_queries")
                 .findWhenNeeded(this);
-
-        // Fields column
-        protected final WebElement fieldsSelectionPanel = Locator.byClass("filter-modal__col_fields")
-                .findWhenNeeded(this);
-
-        // Values column
-        protected final WebElement filterPanel = Locator.byClass("filter-modal__col_filter_exp")
-                .findWhenNeeded(this);
-        protected final Tabs filterTabs = new Tabs.TabsFinder(getDriver()).refindWhenNeeded(filterPanel);
-        protected final FilterExpressionPanel filterExpressionPanel =
-                new FilterExpressionPanel.FilterExpressionPanelFinder(getDriver()).refindWhenNeeded(filterPanel);
-        protected final FilterFacetedPanel filterFacetedPanel =
-                new FilterFacetedPanel.FilterFacetedPanelFinder(getDriver()).refindWhenNeeded(filterPanel);
-
-        protected final WebElement submitButton = Locator.css(".modal-footer .btn-success").findWhenNeeded(this);
     }
 
 }
